requirements.txt
streamlit
opencv-python-headless
onnxruntime==1.17.3
numpy==1.26.4
pillow
requests
